requirements.txt
flask==2.3.2
gunicorn==20.1.0
jsonify==0.5
pymongo==4.3.3
python-dotenv==1.0.0